Revert "[REFACTORING] Split PullRequestUpdater non-/codeflow (#5720)"
This reverts commit 95ebc7cf56558455c877f3f7f687d86b73100434.
