+++ src/ProductConstructionService/ProductConstructionService.DependencyFlow/Model/PullRequestUpdaterId.cs
@@ -6,12 +6,28 @@ using Maestro.Data.Models;
 
 namespace ProductConstructionService.DependencyFlow.Model;
 
+public class UpdaterId
+{
+    public string Id { get; }
+
+    public UpdaterId(string id)
+    {
+        Id = id;
+    }
+
+    public override string ToString() => Id.ToString();
+
+    public override bool Equals(object? obj) => Id.Equals((obj as UpdaterId)?.Id);
+
+    public override int GetHashCode() => Id.GetHashCode();
+}
+
 public class NonBatchedPullRequestUpdaterId : PullRequestUpdaterId
 {
     public Guid SubscriptionId { get; }
 
-    public NonBatchedPullRequestUpdaterId(Guid subscriptionId, bool isCodeFlow)
-        : base(subscriptionId.ToString(), isCodeFlow)
+    public NonBatchedPullRequestUpdaterId(Guid subscriptionId)
+        : base(subscriptionId.ToString())
     {
         SubscriptionId = subscriptionId;
     }
@@ -22,35 +38,30 @@ public class BatchedPullRequestUpdaterId : PullRequestUpdaterId
     public string Repository { get; }
     public string Branch { get; }
 
-    public BatchedPullRequestUpdaterId(string repository, string branch, bool isCodeFlow)
-        : base(Encode(repository) + ":" + Encode(branch), isCodeFlow)
+    public BatchedPullRequestUpdaterId(string repository, string branch)
+        : base(Encode(repository) + ":" + Encode(branch))
     {
         Repository = repository;
         Branch = branch;
     }
 }
 
-public abstract class PullRequestUpdaterId
+public abstract class PullRequestUpdaterId : UpdaterId
 {
-    public string Id { get; }
-
-    public bool IsCodeFlow { get; }
-
-    protected PullRequestUpdaterId(string id, bool isCodeFlow)
+    protected PullRequestUpdaterId(string id)
+        : base(id)
     {
-        Id = id;
-        IsCodeFlow = isCodeFlow;
     }
 
     /// <summary>
     ///     Parses an <see cref="UpdaterId" /> created by <see cref="Create(string, string)" /> into the (repository, branch)
     ///     pair that created it.
     /// </summary>
-    public static PullRequestUpdaterId Parse(string id, bool isCodeFlow)
+    public static PullRequestUpdaterId Parse(string id)
     {
         if (Guid.TryParse(id, out var guid))
         {
-            return new NonBatchedPullRequestUpdaterId(guid, isCodeFlow);
+            return new NonBatchedPullRequestUpdaterId(guid);
         }
 
         var colonIndex = id.IndexOf(':');
@@ -61,7 +72,7 @@ public abstract class PullRequestUpdaterId
 
         var repository = Decode(id.Substring(0, colonIndex));
         var branch = Decode(id.Substring(colonIndex + 1));
-        return new BatchedPullRequestUpdaterId(repository, branch, isCodeFlow);
+        return new BatchedPullRequestUpdaterId(repository, branch);
     }
 
     protected static string Encode(string repository)
@@ -79,13 +90,7 @@ public abstract class PullRequestUpdaterId
         ArgumentNullException.ThrowIfNull(subscription);
 
         return subscription.PolicyObject.Batchable
-            ? new BatchedPullRequestUpdaterId(subscription.TargetRepository, subscription.TargetBranch, subscription.SourceEnabled)
-            : new NonBatchedPullRequestUpdaterId(subscription.Id, subscription.SourceEnabled);
+            ? new BatchedPullRequestUpdaterId(subscription.TargetRepository, subscription.TargetBranch)
+            : new NonBatchedPullRequestUpdaterId(subscription.Id);
     }
-
-    public override string ToString() => Id.ToString();
-
-    public override bool Equals(object? obj) => Id.Equals((obj as PullRequestUpdaterId)?.Id);
-
-    public override int GetHashCode() => Id.GetHashCode();
 }
