+++ src/ProductConstructionService/ProductConstructionService.DependencyFlow/NonBatchedPullRequestUpdater.cs
@@ -5,23 +5,24 @@ using Maestro.Data;
 using Maestro.Data.Models;
 using Maestro.DataProviders;
 using Microsoft.DotNet.DarcLib;
+using Microsoft.DotNet.DarcLib.VirtualMonoRepo;
 using Microsoft.Extensions.Logging;
 using ProductConstructionService.Common;
 using ProductConstructionService.DependencyFlow.Model;
 using ProductConstructionService.WorkItems;
 
-namespace ProductConstructionService.DependencyFlow.PullRequestUpdaters;
+namespace ProductConstructionService.DependencyFlow;
 
-internal class NonBatchedDependencyPullRequestUpdater : DependencyPullRequestUpdater
+internal class NonBatchedPullRequestUpdater : PullRequestUpdater
 {
     private readonly Lazy<Task<Subscription?>> _lazySubscription;
     private readonly NonBatchedPullRequestUpdaterId _id;
     private readonly BuildAssetRegistryContext _context;
-    private readonly ILogger<NonBatchedDependencyPullRequestUpdater> _logger;
+    private readonly ILogger<NonBatchedPullRequestUpdater> _logger;
     private readonly ICommentCollector _commentCollector;
     private readonly IPullRequestCommentBuilder _commentBuilder;
 
-    public NonBatchedDependencyPullRequestUpdater(
+    public NonBatchedPullRequestUpdater(
         NonBatchedPullRequestUpdaterId id,
         IMergePolicyEvaluator mergePolicyEvaluator,
         BuildAssetRegistryContext context,
@@ -32,11 +33,16 @@ internal class NonBatchedDependencyPullRequestUpdater : DependencyPullRequestUpd
         IRedisCacheFactory cacheFactory,
         IReminderManagerFactory reminderManagerFactory,
         ISqlBarClient sqlClient,
+        ILocalLibGit2Client gitClient,
+        IVmrInfo vmrInfo,
+        IPcsVmrForwardFlower vmrForwardFlower,
+        IPcsVmrBackFlower vmrBackFlower,
+        ITelemetryRecorder telemetryRecorder,
+        ILogger<NonBatchedPullRequestUpdater> logger,
         ICommentCollector commentCollector,
         IPullRequestCommenter pullRequestCommenter,
         IPullRequestCommentBuilder commentBuilder,
-        IFeatureFlagService featureFlagService,
-        ILogger<NonBatchedDependencyPullRequestUpdater> logger)
+        IFeatureFlagService featureFlagService)
         : base(
             id,
             mergePolicyEvaluator,
@@ -48,7 +54,13 @@ internal class NonBatchedDependencyPullRequestUpdater : DependencyPullRequestUpd
             cacheFactory,
             reminderManagerFactory,
             sqlClient,
+            gitClient,
+            vmrInfo,
+            vmrForwardFlower,
+            vmrBackFlower,
+            telemetryRecorder,
             logger,
+            commentCollector,
             pullRequestCommenter,
             featureFlagService)
     {
@@ -60,20 +72,26 @@ internal class NonBatchedDependencyPullRequestUpdater : DependencyPullRequestUpd
         _commentBuilder = commentBuilder;
     }
 
+    public Guid SubscriptionId => _id.SubscriptionId;
+
     private async Task<Subscription?> RetrieveSubscription()
     {
-        Subscription? subscription = await _context.Subscriptions.FindAsync(_id.SubscriptionId);
+        Subscription? subscription = await _context.Subscriptions.FindAsync(SubscriptionId);
 
         // This can mainly happen during E2E tests where we delete a subscription
         // while some PRs have just been closed and there's a reminder on those still
         if (subscription == null)
         {
-            _logger.LogWarning(
-                "Failed to find a subscription {subscriptionId}. " +
-                "Possibly it was deleted while an existing PR is still tracked. Untracking PR...",
-                _id.SubscriptionId);
+            _logger.LogInformation(
+                $"Failed to find a subscription {SubscriptionId}. " +
+                "Possibly it was deleted while an existing PR is still tracked. Untracking PR...");
 
-            await ClearAllStateAsync(clearPendingUpdates: true);
+            // We don't know if the subscription was a code flow one, so just unset both
+            await _pullRequestState.TryDeleteAsync();
+            await _pullRequestCheckReminders.UnsetReminderAsync(isCodeFlow: true);
+            await _pullRequestCheckReminders.UnsetReminderAsync(isCodeFlow: false);
+            await _pullRequestUpdateReminders.UnsetReminderAsync(isCodeFlow: true);
+            await _pullRequestUpdateReminders.UnsetReminderAsync(isCodeFlow: false);
             return null;
         }
 
@@ -97,7 +115,7 @@ internal class NonBatchedDependencyPullRequestUpdater : DependencyPullRequestUpd
     protected override async Task<(string repository, string branch)> GetTargetAsync()
     {
         Subscription subscription = await GetSubscription()
-            ?? throw new SubscriptionException($"Subscription '{_id.SubscriptionId}' was not found...");
+            ?? throw new SubscriptionException($"Subscription '{SubscriptionId}' was not found...");
         return (subscription.TargetRepository, subscription.TargetBranch);
     }
 
@@ -105,5 +123,20 @@ internal class NonBatchedDependencyPullRequestUpdater : DependencyPullRequestUpd
     {
         Subscription? subscription = await GetSubscription();
         return subscription?.PolicyObject?.MergePolicies ?? [];
+    }
+
+    protected override async Task<bool> CheckInProgressPullRequestAsync(
+        InProgressPullRequest pullRequestCheck,
+        bool isCodeFlow)
+    {
+        Subscription? subscription = await GetSubscription();
+        if (subscription == null)
+        {
+            // If the subscription was deleted during tests (a frequent occurrence when we delete subscriptions at the end),
+            // we don't want to report this as a failure. For real PRs, it might be good to learn about this. 
+            return pullRequestCheck.Url?.Contains("maestro-auth-test") ?? false;
+        }
+
+        return await base.CheckInProgressPullRequestAsync(pullRequestCheck, isCodeFlow);
     }
 }
