+++ test/ProductConstructionService.DependencyFlow.Tests/PullRequestBuilderTests.cs
@@ -12,7 +12,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Moq;
 using NUnit.Framework;
 using ProductConstructionService.DependencyFlow.Model;
-using ProductConstructionService.DependencyFlow.PullRequestUpdaters;
 using ProductConstructionService.DependencyFlow.WorkItems;
 
 using Asset = Microsoft.DotNet.ProductConstructionService.Client.Models.Asset;
