+++ test/ProductConstructionService.DependencyFlow.Tests/SubscriptionUpdaterTests.cs
@@ -8,7 +8,6 @@ using Microsoft.VisualStudio.Services.Common;
 using Moq;
 using NUnit.Framework;
 using ProductConstructionService.DependencyFlow.Model;
-using ProductConstructionService.DependencyFlow.PullRequestUpdaters;
 using Asset = ProductConstructionService.DependencyFlow.Model.Asset;
 
 namespace ProductConstructionService.DependencyFlow.Tests;
@@ -16,7 +15,7 @@ namespace ProductConstructionService.DependencyFlow.Tests;
 [TestFixture, NonParallelizable]
 internal class SubscriptionUpdaterTests : SubscriptionOrPullRequestUpdaterTests
 {
-    protected Dictionary<PullRequestUpdaterId, Mock<IPullRequestUpdater>> PullRequestUpdaters { get; private set; } = [];
+    protected Dictionary<UpdaterId, Mock<IPullRequestUpdater>> PullRequestUpdaters { get; private set; } = [];
 
     [SetUp]
     public void SubscriptionUpdaterTests_SetUp()
@@ -54,7 +53,7 @@ internal class SubscriptionUpdaterTests : SubscriptionOrPullRequestUpdaterTests
             });
     }
 
-    private void ThenUpdateAssetsAsyncShouldHaveBeenCalled(PullRequestUpdaterId forUpdater, Build withBuild)
+    private void ThenUpdateAssetsAsyncShouldHaveBeenCalled(UpdaterId forUpdater, Build withBuild)
     {
         var updatedAssets = new List<List<Asset>>();
         PullRequestUpdaters.Should().ContainKey(forUpdater)
@@ -81,7 +80,7 @@ internal class SubscriptionUpdaterTests : SubscriptionOrPullRequestUpdaterTests
 
         await WhenUpdateAsyncIsCalled(Subscription, b);
         ThenUpdateAssetsAsyncShouldHaveBeenCalled(
-            new BatchedPullRequestUpdaterId(Subscription.TargetRepository, Subscription.TargetBranch, Subscription.SourceEnabled),
+            new BatchedPullRequestUpdaterId(Subscription.TargetRepository, Subscription.TargetBranch),
             b);
     }
 
@@ -98,6 +97,6 @@ internal class SubscriptionUpdaterTests : SubscriptionOrPullRequestUpdaterTests
         Build b = GivenANewBuild(true);
 
         await WhenUpdateAsyncIsCalled(Subscription, b);
-        ThenUpdateAssetsAsyncShouldHaveBeenCalled(new NonBatchedPullRequestUpdaterId(Subscription.Id, Subscription.SourceEnabled), b);
+        ThenUpdateAssetsAsyncShouldHaveBeenCalled(new NonBatchedPullRequestUpdaterId(Subscription.Id), b);
     }
 }
